requirements.txt
pandas==1.5.3
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==0.21.1
